refactor(bridge-scene): share the item-urn and token-urn helpers
wearables and emotes each had their own copy of the tokenId strip and the
owned-catalog token-urn builder. Move both to a urns module. The emotes copy
only stripped collections-v2; the shared one also handles v1, which no emote
uses, so no behaviour change.
